+++ CHANGELOG.md
@@ -11,6 +11,7 @@ If you are upgrading from v0.16.x, replace the binary (or run `docker pull`). If
 ## Changed
 
 ## Fixed
+- Directory: Invalidate caches when group memberships change on an external directory.
 - Log viewer: All events show as `INFO`.
 
 ## [0.16.2] - 2026-04-28